Basic skeleton of the databases
This is a very basic skeleton. Still need to tweak the table fields, and create the following APIs: create new recipe, get recipes by user (using email as an ID), get all recipes, update recipe, delete recipe.

Unsure what is wanted for the steps needed for a recipe, may or may not add it depending on the time leftover after classes and internship.